A bunch of minor fixes for the API

+++ sugarcrm/include/api/RelateRecordApi.php
@@ -79,6 +79,9 @@ class RelateRecordApi extends ModuleApi {
      * @return array Two elements: The link name, and the SugarBean of the far end
      */
     protected function checkRelatedSecurity(ServiceBase $api, $args, SugarBean $primaryBean, $securityTypeLocal='view', $securityTypeRemote='view') {
+        if ( empty($primaryBean) ) {
+            throw new SugarApiExceptionNotFound('Could not find the primary bean');
+        }
         if ( ! $primaryBean->ACLAccess($securityTypeLocal) ) {
             throw new SugarApiExceptionNotAuthorized('No access to '.$securityTypeLocal.' records for module: '.$args['module']);
         }
@@ -156,6 +159,10 @@ class RelateRecordApi extends ModuleApi {
         list($linkName, $relatedBean) = $this->checkRelatedSecurity($api, $args, $primaryBean, 'view','view');
 
         $relatedBean->retrieve($args['remote_id']);
+        if ( empty($relatedBean->id) ) {
+            // Retrieve failed, probably doesn't have permissions
+            throw new SugarApiExceptionNotFound('Could not find the related bean');
+        }
 
         return $this->formatBean($api, $args, $relatedBean);
         
@@ -180,9 +187,11 @@ class RelateRecordApi extends ModuleApi {
         SugarRelationship::resaveRelatedBeans();
 
         // Reload the related record so that it has all the latest data similar to what we do in the create and update
-        // This needs to be done after the resaveRelatedBeans not before.  This will allow the relationship to stick.
         $relatedBean->retrieve($id);
-
+        if ( empty($relatedBean->id) ) {
+            // Retrieve failed, probably doesn't have permissions
+            throw new SugarApiExceptionNotFound('Could not find the related bean');
+        }
         return $this->formatNearAndFarRecords($api,$args,$primaryBean,$relatedBean,$linkName,$relatedData);
     }
 
@@ -192,6 +201,10 @@ class RelateRecordApi extends ModuleApi {
         list($linkName, $relatedBean) = $this->checkRelatedSecurity($api, $args, $primaryBean, 'view','view');
 
         $relatedBean->retrieve($args['remote_id']);
+        if ( empty($relatedBean->id) ) {
+            // Retrieve failed, probably doesn't have permissions
+            throw new SugarApiExceptionNotFound('Could not find the related bean');
+        }
         
         $relatedData = $this->getRelatedFields($api, $args, $primaryBean, $linkName);
 
@@ -207,6 +220,10 @@ class RelateRecordApi extends ModuleApi {
         list($linkName, $relatedBean) = $this->checkRelatedSecurity($api, $args, $primaryBean, 'view','edit');
 
         $relatedBean->retrieve($args['remote_id']);
+        if ( empty($relatedBean->id) ) {
+            // Retrieve failed, probably doesn't have permissions
+            throw new SugarApiExceptionNotFound('Could not find the related bean');
+        }
 
         $id = $this->updateBean($relatedBean, $api, $args);
 
@@ -226,6 +243,10 @@ class RelateRecordApi extends ModuleApi {
         list($linkName, $relatedBean) = $this->checkRelatedSecurity($api, $args, $primaryBean, 'view','view');
 
         $relatedBean->retrieve($args['remote_id']);
+        if ( empty($relatedBean->id) ) {
+            // Retrieve failed, probably doesn't have permissions
+            throw new SugarApiExceptionNotFound('Could not find the related bean');
+        }
 
         $primaryBean->$linkName->delete($primaryBean->id,$relatedBean);
